requirements.txt
asgiref==3.8.1
coverage==7.6.1
Django==4.2
django-timezone-field==7.0
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.1
python-dotenv==1.0.1
pillow==10.4.0
psycopg2==2.9.9
redis==5.0.8
amqp==5.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
kombu==5.4.1
prompt-toolkit==3.0.47
python-dateutil==2.9.0.post0
six==1.16.0
vine==5.1.0
wcwidth==0.2.13
